requirements_api.txt
# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyngrok>=7.0.0
python-multipart>=0.0.6

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy<2

# OCR dependencies (from requirement.txt)
einops==0.2.0
paddlepaddle
pyclipper
lmdb

# Additional dependencies
requests>=2.31.0
